fix(Climbing): inverted winch limit switch and logic for move_winch
Removed the inversion from `MoveWinch.execute` for simplicity of control
management.

The speed is currently coming through as negative for upwards and
positive for downwards. Inverted the `move_winch` method logic to
handle the limit switch inversion.

Add the value of the limit switch to the smart dashboard.

Also added the value of `is_retracted` to the smart dashboard for
visibility.

Renamed smart dashboard fields to "Winch"

TODO:
Make move_winch read better while supporting inversion.