feat(ControlPanel, ColorSensorV3): new commands and Color Sensing
Added a command to cancel the spin of the Control Panel.

Added a command to run the spin of the Control Panel for an arbitrary amount of time.

Added button press events for cancel and timed spins.

Added a mapping of Control Panel detected color to scored color. This isn't accurate
right now because of the possibility that multiple colors might be scored from a single
detected color.