Fix/shooter (#29)
* feat(ControlPanelArm): add ControlPanelArm Subsystem

Adding a ControlPanelArm subsystem to prevent requirement conflicts in Control Panel.

* feat(ControlPanelArm): fix set inverted and add subsystems.ini config

Added `ControlPanelArm` configuration section.

* fix(ControlPanelArm): reorder configuration

* fix(ControlPanel, ControlPanelArm): add `ControlPanelArm` subsystem to robot

Remove the control panel arm from the Control Panel class.

Switched `LowerArm` and `RaiseArm` to point to the extend function of
`ControlPanelArm`.

* fix(ControlPanelArm): remove color profile load

* fix(ControlPanel): set max speed to 1 and removed solenoid

* fix(Shooter): shooter sorta functional

Added command to stop the shooter from shooting. `DoNothing` was preventing
the shooter from operating. It didn't take the robot or the require the `Shooter`
subsystem.

Added a reverse command for the shooter, as well as setting it to the left bumper.