Feature/control panel spin time (#31)
* feat(ControlPanelArm): add ControlPanelArm Subsystem

Adding a ControlPanelArm subsystem to prevent requirement conflicts in Control Panel.

* feat(ControlPanelArm): fix set inverted and add subsystems.ini config

Added `ControlPanelArm` configuration section.

* fix(ControlPanelArm): reorder configuration

* fix(ControlPanel, ControlPanelArm): add `ControlPanelArm` subsystem to robot

Remove the control panel arm from the Control Panel class.

Switched `LowerArm` and `RaiseArm` to point to the extend function of
`ControlPanelArm`.

* fix(ControlPanelArm): remove color profile load

* fix(ControlPanel): set max speed to 1 and removed solenoid

* fix(Shooter): shooter sorta functional

Added command to stop the shooter from shooting. `DoNothing` was preventing
the shooter from operating. It didn't take the robot or the require the `Shooter`
subsystem.

Added a reverse command for the shooter, as well as setting it to the left bumper.

* fix(ControlPanel, ColorSensorV3): Control Panel color sensor working

Added `get_current_color` to move call for Control Panel.

Renamed color_matcher.matchClosesColor method.

Added confidence interval to `matchClosestColor`.

Added `ControlPanel.PanelColor` detected by ColorMatch
to smartdashboard.

* feat(ColorSensorV3): set color value from read sensor values

Set target practice colors based on values seen from practice.

Used swatches from the robotics kit.

* feat(ControlPanel, ColorSensorV3): new commands and Color Sensing

Added a command to cancel the spin of the Control Panel.

Added a command to run the spin of the Control Panel for an arbitrary amount of time.

Added button press events for cancel and timed spins.

Added a mapping of Control Panel detected color to scored color. This isn't accurate
right now because of the possibility that multiple colors might be scored from a single
detected color.

Co-authored-by: Aaron <account+github@aarondegrow.com>